[ refactor ] Move some unwanted functions back
